Update Object class its use
Now Objects gravity effect other
objects. To demonstrate, the
default orbit demo now has an
asteroid added.

+++ src/Player.cpp
@@ -102,7 +102,7 @@ Player :: ~Player()
 
 }
 
-void Player :: setRotation_rate(int rotRate)
+void Player :: setRotation_rate(float rotRate)
 {
 	//rotationRate = rotRate;
 	sprite.setRotationRate(rotRate);
